@@ -1,17 +1,26 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <PreserveCompilationContext>true</PreserveCompilationContext>
     <AssemblyName>PartsUnlimitedWebsite</AssemblyName>
     <AssetTargetFallback>$(AssetTargetFallback);portable-net45+win8+wp8+wpa81;</AssetTargetFallback>
+	<AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
     <UserSecretsId>aspnet-PartsUnlimitedWebsite-88B5164A-732F-489C-B25E-AC693ECEB5FB</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>	
+    
+    <PackageReference Include="Microsoft.AspNetCore.Diagnostics.EntityFrameworkCore" Version="3.1.1" />	
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.OpenIdConnect" Version="3.1.1" />	
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.Facebook" Version="3.1.1" />	
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.Twitter" Version="3.1.1" />	
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.Google" Version="3.1.1" />	
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.MicrosoftAccount" Version="3.1.1" />	
 	<PackageReference Include="Azure.Storage.Queues" Version="12.6.0" />	
-	<PackageReference Include="Microsoft.AspNetCore.All" />	
 	<PackageReference Include="Microsoft.DotNet.InternalAbstractions" Version="1.0.500-preview2-1-003177" />	
+	<PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.19" />	
+	<PackageReference Include="Microsoft.EntityFrameworkCore.Abstractions" Version="3.1.19" />	
 	<PackageReference Include="System.Configuration.ConfigurationManager" Version="5.0.0" />	
 
  </ItemGroup>
@@ -39,7 +48,7 @@
   <Target Name="PrecompileScript" BeforeTargets="BeforeBuild">
     <Exec Command="npm install --python=python2.7" />
     <Exec Command="npm config set python python2.7" />
-    <Exec Command="npm install -g grunt-cli" />
+    <Exec Command="npm install -g grunt-cli --force" />
     <Exec Command="grunt" />
   </Target>
 
