@@ -1,10 +1,10 @@
 ﻿using System;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using PartsUnlimited.Models;
+using Microsoft.Extensions.Hosting;
 
 namespace PartsUnlimited
 {
@@ -12,7 +12,7 @@ namespace PartsUnlimited
     {
         public static void Main(string[] args)
         {
-            var host = BuildWebHost(args);
+            var host = BuildWebHost(args).Build();
 
             using (var scope = host.Services.CreateScope())
             {
@@ -33,14 +33,17 @@ namespace PartsUnlimited
             host.Run();
         }
 
-        public static IWebHost BuildWebHost(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>()
-                .ConfigureAppConfiguration((hostContext, config) =>
+        public static IHostBuilder BuildWebHost(string[] args) =>
+            Host.CreateDefaultBuilder(args).ConfigureAppConfiguration((hostContext, config) =>
                 {
                     config.AddJsonFile("config.json", optional: true);
                     config.AddEnvironmentVariables();
                 })
-                .Build();
+                .ConfigureWebHostDefaults(webBuilder =>
+                {
+                    webBuilder.UseStartup<Startup>();
+                });   
+        
+       
     }
 }
