@@ -7,22 +7,21 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.EntityFrameworkCore.Extensions.Internal;
 
 namespace PartsUnlimited.Queries
 {
     public class RaincheckQuery : IRaincheckQuery
     {
-        private readonly IPartsUnlimitedContext _context;
+        private readonly PartsUnlimitedContext _context;
 
-        public RaincheckQuery(IPartsUnlimitedContext context)
+        public RaincheckQuery(PartsUnlimitedContext context)
         {
             _context = context;
         }
 
         public async Task<IEnumerable<Raincheck>> GetAllAsync()
         {
-            var rainchecks = await _context.RainChecks.AsAsyncEnumerable().ToList();
+            var rainchecks = _context.RainChecks.ToList();
 
             foreach (var raincheck in rainchecks)
             {
@@ -34,7 +33,7 @@ namespace PartsUnlimited.Queries
 
         public async Task<Raincheck> FindAsync(int id)
         {
-            var raincheck = await _context.RainChecks.AsAsyncEnumerable().FirstOrDefault(r => r.RaincheckId == id);
+            var raincheck = _context.RainChecks.FirstOrDefault(r => r.RaincheckId == id);
 
             if (raincheck == null)
             {
@@ -60,9 +59,9 @@ namespace PartsUnlimited.Queries
         /// </summary>
         private async Task FillRaincheckValuesAsync(Raincheck raincheck)
         {
-            raincheck.IssuerStore = await _context.Stores.AsAsyncEnumerable().First(s => s.StoreId == raincheck.StoreId);
-            raincheck.Product = await _context.Products.AsAsyncEnumerable().First(p => p.ProductId == raincheck.ProductId);
-            raincheck.Product.Category = await _context.Categories.AsAsyncEnumerable().First(c => c.CategoryId == raincheck.Product.CategoryId);
+            raincheck.IssuerStore =  _context.Stores.First(s => s.StoreId == raincheck.StoreId);
+            raincheck.Product = _context.Products.First(p => p.ProductId == raincheck.ProductId);
+            raincheck.Product.Category = _context.Categories.ToList().First(c => c.CategoryId == raincheck.Product.CategoryId);
         }
     }
 }
